Improved edge case handling of move command/keys. Removed a number of compiler warnings
git-svn-id: https://vimpc.svn.sourceforge.net/svnroot/vimpc/trunk@292 fc90b29e-3040-4360-8be2-b397c2834534

+++ src/buffer/library.hpp
@@ -117,7 +117,7 @@ namespace Mpc
             parent_->AddPartial();
          }
 
-         if ((parent_ != NULL) && ((childrenInPlaylist_ == children_.size()) || (type_ == Mpc::SongType)))
+         if ((parent_ != NULL) && ((childrenInPlaylist_ == static_cast<int32_t>(children_.size())) || (type_ == Mpc::SongType)))
          {
             parent_->AddedToPlaylist();
          }
@@ -125,7 +125,7 @@ namespace Mpc
 
       void RemovedFromPlaylist()
       {
-         if ((parent_ != NULL) && ((childrenInPlaylist_ == children_.size()) || (type_ == Mpc::SongType)))
+         if ((parent_ != NULL) && ((childrenInPlaylist_ == static_cast<int32_t>(children_.size())) || (type_ == Mpc::SongType)))
          {
             parent_->RemovedFromPlaylist();
          }
